requirements.txt
faiss-cpu
numpy
tqdm
sentence-transformers
ollama
wikipedia-api
